Reorganize events API into events/views/api/ and events/serializers/
Move EventListAPIView and FutureEventListAPIView to events/views/api/event_list.py,
EventCreateAPIView to events/views/api/event_create.py, and consolidate
PlaceSerializer + EventListSerializer (renamed from EventSerializer) into
events/serializers/event.py — deleting the events/api/ package entirely.
Tests and mock patch targets updated to match the new paths.
