+++ events/serializers/__init__.py
@@ -1,3 +1,3 @@
-from events.serializers.event import EventWriteSerializer
+from events.serializers.event import EventListSerializer, EventWriteSerializer
 
-__all__ = ['EventWriteSerializer']
+__all__ = ['EventListSerializer', 'EventWriteSerializer']
