+++ events/tests/views/api/test_event_list.py
