+++ events/views/api/event_list.py
@@ -1,0 +1,23 @@
+from django_filters.rest_framework import DjangoFilterBackend
+from rest_framework.filters import OrderingFilter
+from rest_framework.generics import ListAPIView
+
+from events.models import Event
+from events.serializers.event import EventListSerializer
+
+
+class EventListAPIView(ListAPIView):
+    serializer_class = EventListSerializer
+    filter_backends = [DjangoFilterBackend, OrderingFilter]
+    filterset_fields = ['place__city__slug']
+    ordering_fields = ['event_date']
+
+    def get_queryset(self):
+        return Event.objects.published().order_by('event_date')
+
+
+class FutureEventListAPIView(EventListAPIView):
+
+    def get_queryset(self):
+        queryset = super().get_queryset().filter(is_hidden=False)
+        return queryset.future().select_related('place__city')
